+++ pcap4j-core/src/main/java/org/pcap4j/core/RawPacketListener.java
@@ -1,0 +1,23 @@
+/*_##########################################################################
+  _##
+  _##  Copyright (C) 2015  Kaito Yamada
+  _##
+  _##########################################################################
+*/
+
+package org.pcap4j.core;
+
+
+/**
+ * @author Kaito Yamada
+ * @since pcap4j 1.3.1
+ */
+public interface RawPacketListener {
+
+  /**
+   *
+   * @param packet
+   */
+  public void gotPacket(byte[] packet);
+
+}